requirements.txt
flask==2.2.3
sentence-transformers==2.2.2
transformers==4.28.1
huggingface-hub==0.14.1
python-dotenv==1.0.0
requests==2.28.2